# swagger.json
{
   "swagger":"2.0",
   "info":{
      "description":"Api Documentation",
      "version":"1.0",
      "title":"Api Documentation",
      "termsOfService":"urn:tos",
      "contact":{

      },
      "license":{
         "name":"Apache 2.0",
         "url":"http://www.apache.org/licenses/LICENSE-2.0"
      }
   },
   "host":"localhost:9010",
   "basePath":"/",
...
   "paths":{
...
      "/metric-type/":{
         "get":{
            "tags":[
               "metric-type-controller"
            ],
            "summary":"getAllMetricType",
            "operationId":"getAllMetricTypeUsingGET",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/MetricTypeDTO"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "metric-type-controller"
            ],
            "summary":"createMetricType",
            "operationId":"createMetricTypeUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"metricType",
                  "description":"metricType",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/MetricTypeDTO"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/MetricTypeDTO"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/metric-type/metric/":{
         "get":{
            "tags":[
               "metric-type-controller"
            ],
            "summary":"getAllMetricTypeMetrics",
            "operationId":"getAllMetricTypeMetricsUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"direction",
                  "in":"query",
                  "description":"direction",
                  "required":false,
                  "type":"string",
                  "default":"DESC"
               },
               {
                  "name":"page",
                  "in":"query",
                  "description":"page",
                  "required":false,
                  "type":"integer",
                  "default":0,
                  "format":"int32"
               },
               {
                  "name":"size",
                  "in":"query",
                  "description":"size",
                  "required":false,
                  "type":"integer",
                  "default":10,
                  "format":"int32"
               },
               {
                  "name":"timeGranularity",
                  "in":"query",
                  "description":"timeGranularity",
                  "required":false,
                  "type":"integer",
                  "format":"int64"
               },
               {
                  "name":"unixTimeFrom",
                  "in":"query",
                  "description":"unixTimeFrom",
                  "required":false,
                  "type":"integer",
                  "format":"int64"
               },
               {
                  "name":"unixTimeTo",
                  "in":"query",
                  "description":"unixTimeTo",
                  "required":false,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/MetricTypeValuesDTO"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/metric-type/{id}":{
         "put":{
            "tags":[
               "metric-type-controller"
            ],
            "summary":"updateMetricType",
            "operationId":"updateMetricTypeUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"string"
               },
               {
                  "in":"body",
                  "name":"metricType",
                  "description":"metricType",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/MetricTypeDTO"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/MetricTypeDTO"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "metric-type-controller"
            ],
            "summary":"deleteMetricType",
            "operationId":"deleteMetricTypeUsingDELETE",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      },
      "/metric-type/{id}/":{
         "get":{
            "tags":[
               "metric-type-controller"
            ],
            "summary":"getMetricTypeById",
            "operationId":"getMetricTypeByIdUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/MetricTypeDTO"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/metric-type/{metricTypeId}/metric/":{
         "get":{
            "tags":[
               "metric-controller"
            ],
            "summary":"getAllMetricsForMetricType",
            "operationId":"getAllMetricsForMetricTypeUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"direction",
                  "in":"query",
                  "description":"direction",
                  "required":false,
                  "type":"string",
                  "default":"DESC"
               },
               {
                  "name":"metricTypeId",
                  "in":"path",
                  "description":"metricTypeId",
                  "required":true,
                  "type":"string"
               },
               {
                  "name":"page",
                  "in":"query",
                  "description":"page",
                  "required":false,
                  "type":"integer",
                  "default":0,
                  "format":"int32"
               },
               {
                  "name":"size",
                  "in":"query",
                  "description":"size",
                  "required":false,
                  "type":"integer",
                  "default":10,
                  "format":"int32"
               },
               {
                  "name":"timeGranularity",
                  "in":"query",
                  "description":"timeGranularity",
                  "required":false,
                  "type":"integer",
                  "format":"int64"
               },
               {
                  "name":"unixTimeFrom",
                  "in":"query",
                  "description":"unixTimeFrom",
                  "required":false,
                  "type":"integer",
                  "format":"int64"
               },
               {
                  "name":"unixTimeTo",
                  "in":"query",
                  "description":"unixTimeTo",
                  "required":false,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Page«MetricDTO»"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "metric-controller"
            ],
            "summary":"createMetric",
            "operationId":"createMetricUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"metricDTO",
                  "description":"metricDTO",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/MetricDTO"
                  }
               },
               {
                  "name":"metricTypeId",
                  "in":"path",
                  "description":"metricTypeId",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/MetricDTO"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/metric-type/{metricTypeId}/metric/{metricId}/":{
         "put":{
            "tags":[
               "metric-controller"
            ],
            "summary":"updateMetric",
            "operationId":"updateMetricUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"metricDTO",
                  "description":"metricDTO",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/MetricDTO"
                  }
               },
               {
                  "name":"metricId",
                  "in":"path",
                  "description":"metricId",
                  "required":true,
                  "type":"string"
               },
               {
                  "name":"metricTypeId",
                  "in":"path",
                  "description":"metricTypeId",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/MetricDTO"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "metric-controller"
            ],
            "summary":"deleteMetric",
            "operationId":"deleteMetricUsingDELETE",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"metricId",
                  "in":"path",
                  "description":"metricId",
                  "required":true,
                  "type":"string"
               },
               {
                  "name":"metricTypeId",
                  "in":"path",
                  "description":"metricTypeId",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      }
   },
   "definitions":{
      "MetricDTO":{
         "type":"object",
         "properties":{
            "date":{
               "type":"string",
               "format":"date-time"
            },
            "id":{
               "type":"integer",
               "format":"int64"
            },
            "metricType":{
               "type":"string"
            },
            "value":{
               "type":"number",
               "format":"double"
            }
         },
         "title":"MetricDTO"
      },
      "MetricTypeDTO":{
         "type":"object",
         "properties":{
            "id":{
               "type":"integer",
               "format":"int64"
            },
            "informationFormat":{
               "type":"string",
               "enum":[
                  "MB",
                  "PERCENTAGE"
               ]
            },
            "name":{
               "type":"string"
            }
         },
         "title":"MetricTypeDTO"
      },
      "MetricTypeValuesDTO":{
         "type":"object",
         "properties":{
            "id":{
               "type":"integer",
               "format":"int64"
            },
            "informationFormat":{
               "type":"string",
               "enum":[
                  "MB",
                  "PERCENTAGE"
               ]
            },
            "metrics":{
               "$ref":"#/definitions/Page«MetricDTO»"
            },
            "name":{
               "type":"string"
            }
         },
         "title":"MetricTypeValuesDTO"
      },
...
      "Pageable":{
         "type":"object",
         "properties":{
            "offset":{
               "type":"integer",
               "format":"int64"
            },
            "pageNumber":{
               "type":"integer",
               "format":"int32"
            },
            "pageSize":{
               "type":"integer",
               "format":"int32"
            },
            "paged":{
               "type":"boolean"
            },
            "sort":{
               "$ref":"#/definitions/Sort"
            },
            "unpaged":{
               "type":"boolean"
            }
         },
         "title":"Pageable"
      },
      "Page«MetricDTO»":{
         "type":"object",
         "properties":{
            "content":{
               "type":"array",
               "items":{
                  "$ref":"#/definitions/MetricDTO"
               }
            },
            "empty":{
               "type":"boolean"
            },
            "first":{
               "type":"boolean"
            },
            "last":{
               "type":"boolean"
            },
            "number":{
               "type":"integer",
               "format":"int32"
            },
            "numberOfElements":{
               "type":"integer",
               "format":"int32"
            },
            "pageable":{
               "$ref":"#/definitions/Pageable"
            },
            "size":{
               "type":"integer",
               "format":"int32"
            },
            "sort":{
               "$ref":"#/definitions/Sort"
            },
            "totalElements":{
               "type":"integer",
               "format":"int64"
            },
            "totalPages":{
               "type":"integer",
               "format":"int32"
            }
         },
         "title":"Page«MetricDTO»"
      },
      "Sort":{
         "type":"object",
         "properties":{
            "empty":{
               "type":"boolean"
            },
            "sorted":{
               "type":"boolean"
            },
            "unsorted":{
               "type":"boolean"
            }
         },
         "title":"Sort"
      },
...
   }
}